Adds alternate layout, extracts config & vibes
Layout can now place the time at the top or the bottom of the screen. Currently uses the latter layout (time on the bottom) as it's more glanceable. TODO: This option needs to be exposed via configuration.

The clay configuration code has been extracted to its own module. It's not groundbreaking or anything, but it does help clean up main.c. TODO: default values should be read from a Configuration struct.

Adds a vibe_patterns module, but this is still violently buggy.

Assorted cleanup in main.c.