Adds double-tap gesture, initial util libs
Vibrating alerts can be dismissed with a double-tap. This still needs to be made user-configurable.

Generic functionality that is (or should be) useful beyond this watchface has been migrated from under modules/* into utils/*. There's more work to do on this, but it's off to a good start. In particular, about half of the redundant code in the configuration module has been replaced with calls to persist_utils.